chore: move github router dependencies to separate file
